store hashes in BigIntegerFields

+++ comments/static_app_comments/migrations/0001_initial.py
@@ -19,8 +19,8 @@ class Migration(migrations.Migration):
             fields=[
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('article_url', models.URLField(db_index=True)),
-                ('paragraph_hash', models.IntegerField(db_index=True)),
-                ('comment_hash', models.IntegerField()),
+                ('paragraph_hash', models.BigIntegerField(db_index=True)),
+                ('comment_hash', models.BigIntegerField()),
                 ('timestamp', models.DateTimeField(auto_now=True)),
                 ('deleted', models.BooleanField(default=False)),
                 ('text', models.TextField()),
